try my best

+++ tspcc/path.hpp
@@ -17,7 +17,7 @@
 class Path {
 protected:
 	int _size;
-	long _distance;
+	int _distance;
 	int* _nodes;
 	Graph* _graph;
 public:
